Closed the reader, renamed aws to ons_aws to prevent clash
